+++ pkg/deployment/pod/encryption.go
@@ -108,6 +108,13 @@ func GetEncryptionFolderSecretName(name string) string {
 	return n
 }
 
+// GetLicenseRegistryCredentialsSecretName returns the secret name for storing registry credentials used to pull licensed images
+func GetLicenseRegistryCredentialsSecretName(name string) string {
+	n := fmt.Sprintf("%s-rlm", name)
+
+	return n
+}
+
 func IsEncryptionEnabled(i Input) bool {
 	return i.Deployment.RocksDB.IsEncrypted()
 }